move make_axis and _bind_dynamic_axis_sizing into GridPlot
